Fix bugs

+++ counting-language-impl/src/main/java/com/github/mhdirkse/countlang/execution/FunctionCallExpressionCalculation.java
@@ -189,7 +189,7 @@ implements SampleContextBase {
         @Override
         void acceptChildResultDoingStatements(Object value, ExecutionContext context) {
             context.stopFunctionCall(expression);
-            sampleContext.score(((Integer) value).intValue());
+            sampleContext.score(value);
         }
 
         @Override
